requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
Deprecated==1.2.14
exceptiongroup==1.2.2
fastapi==0.115.4
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyGithub==2.5.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
